Make removal of code more conservative (#9)
This is a follow-up to #6, with a fix more in line with what I described here: https://github.com/rollup/rollup-plugin-strip/pull/6#discussion_r149122874. This should work by re-writing with `void 0` in more places rather than outright removing the code, since that sometimes leads to changed behavior or invalid syntax.